ticker logs to console as provider
